+++ scripts/setup-env-gitbash.sh
@@ -10,7 +10,7 @@ BASE_DIR_NAME=$(dirname `which $0`)
 
 GEM_HOME=${BASE_DIR_NAME}/gems/ruby/2.7.0
 GEM_PATH=${BASE_DIR_NAME}/gems/ruby/2.7.0
-PATH=${BASE_DIR_NAME}/ruby/bin:${BASE_DIR_NAME}/gems/ruby/2.7.0/bin:$PATH
+PATH=${BASE_DIR_NAME}/ruby/bin:${BASE_DIR_NAME}/gems/ruby/2.7.0/bin:${BASE_DIR_NAME}/gems/ruby/2.7.0/gems/urbanopt-cli-0.11.1/example_files/python_deps/Miniconda-4.12.0/bin:$PATH
 RUBYLIB=${BASE_DIR_NAME}/OpenStudio/Ruby
 RUBY_DLL_PATH=${BASE_DIR_NAME}/OpenStudio/Ruby
 
@@ -24,4 +24,3 @@ echo "export GEM_PATH=\"${GEM_PATH}\"" >> ~/.env_uo.sh
 echo "export PATH=\"${PATH}\"" >> ~/.env_uo.sh
 echo "export RUBYLIB=\"${RUBYLIB}\"" >> ~/.env_uo.sh
 echo "export RUBY_DLL_PATH=\"${RUBY_DLL_PATH}\"" >> ~/.env_uo.sh
-